refactor: extract gate/plan/proof data cluster into homeboy-gate-contract
The gate, plan, and proof modules form a mutually-dependent pure-data
cluster: plan references HomeboyGateResult + HomeboyProof; gate references
HomeboyPlan; proof references HomeboyGateResult. Because the types depend on
one another they must share a crate, so all three move together DOWN into a
new leaf crate homeboy-gate-contract (deps: serde + serde_json + clap
ValueEnum). Named for gate -- the load-bearing concept (release/verify
gating) that everything else orbits.

What moved (pure data + builders + pure classification helpers):
- gate.rs wholesale (HomeboyGateResult, statuses, collect_plan_gate_results)
- plan.rs wholesale (HomeboyPlan, PlanStep, builders, PlanKind, ...)
- proof.rs data types + builders + the pure helpers proof_scope,
  is_ci_equivalent_gate, gate_status_label, gate_scope_label,
  proof_runner_label, proof_schema, proof_validation_schema

What STAYS in core (behavior with real core coupling):
- proof/validation.rs (validate_proof_value reaches crate::artifact_address
  -> observation)
- proof/loop_spec_validation.rs (LoopSpecValidationProvider behavior hook)

Core re-exports gate + plan from the contract (pub use ... as gate/plan) and
keeps a thin proof module that re-exports the contract proof types plus the
validation glue. All existing paths (core-internal + 10 feature crates + cli)
resolve unchanged. Widened PlanStepBuilder::needs_kind pub(crate)->pub (core
quality.rs calls it cross-crate now).

This unblocks a future gate extraction being clean, and removes the last big
pure-data cluster from the engine.
